Melhorias diversas
